+++ include/simplified/system/interaction/function/pairwise/_G.h
@@ -15,21 +15,21 @@ namespace mysimulator {
 
   template <typename GT,unsigned int DIM>
   void _G_pairwise(
-      Array2D<Float,ArrayKernelName::SSE,__system_vec_type<DIM>::NAME> const& X,
-      Array<UInt> const& ID, Array<Float>& _pre, Array<Float>& _post,
-      Array2D<Float,ArrayKernelName::SSE,__system_vec_type<DIM>::NAME>& _vec,
+      Array2D<Double,ArrayKernelName::SSE,__system_vec_type<DIM>::NAME>const& X,
+      Array<UInt> const& ID, Array<Double>& _pre, Array<Double>& _post,
+      Array2D<Double,ArrayKernelName::SSE,__system_vec_type<DIM>::NAME>& _vec,
       Array<InteractionFunction<GT,DIM>*> const& _neighbor,
       InteractionFuncDataState& _status,
       const InteractionParameter& P, GT const& Geo,
       Array2D<Double,ArrayKernelName::SSE,__system_vec_type<DIM>::NAME>& Grad,
-      float (*_distance_sq)(Array<Float,__system_vec_type<DIM>::NAME>&,
-                            Array<Float,__system_vec_type<DIM>::NAME> const&,
-                            Array<Float,__system_vec_type<DIM>::NAME> const&,
-                            GT const&),
-      void (*_pre_2_post_for_g)(Array<Float> const&,Array<Float>&,
+      double (*_distance_sq)(Array<Double,__system_vec_type<DIM>::NAME>&,
+                             Array<Double,__system_vec_type<DIM>::NAME> const&,
+                             Array<Double,__system_vec_type<DIM>::NAME> const&,
+                             GT const&),
+      void (*_pre_2_post_for_g)(Array<Double> const&,Array<Double>&,
                                 InteractionFuncDataState&,
                                 const InteractionParameter&),
-      void (*_gfunc)(Array<Float> const&,const InteractionParameter&,float*)) {
+      void (*_gfunc)(Array<Double> const&,const InteractionParameter&,double*)){
     assert((bool)X);
     assert((bool)ID);
     assert((bool)_pre);
@@ -47,10 +47,10 @@ namespace mysimulator {
         _distance_sq(_vec[PairwiseVecName::PairBondVecIJ],X[I],X[J],Geo);
       _pre_2_post_for_g(_pre,_post,_status,P);
     }
-    float ef;
+    double ef;
     _gfunc(_post,P,&ef);
     _vec[PairwiseVecName::PairScaledBondVecIJ] = 
-        Float(ef)*_vec[PairwiseVecName::PairBondVecIJ];
+        Double(ef)*_vec[PairwiseVecName::PairBondVecIJ];
     Grad[I]+=_vec[PairwiseVecName::PairScaledBondVecIJ];
     Grad[J]-=_vec[PairwiseVecName::PairScaledBondVecIJ];
   }
