+++ include/simplified/system/interaction/function/pairwise/harmonic/_pre_2_post_for_g.h
@@ -11,7 +11,7 @@
 namespace mysimulator {
 
   void _pre_2_post_for_g_pair_harmonic(
-      Array<Float> const& _pre, Array<Float>& _post,
+      Array<Double> const& _pre, Array<Double>& _post,
       InteractionFuncDataState& _status,const InteractionParameter& P) {
     assert((bool)_pre);
     assert((bool)_post);
